chore(): remove extension lifecycle methods file, update index
